+++ src/NewWords.Api/Services/interfaces/ICurrentUser.cs
@@ -2,7 +2,7 @@ namespace NewWords.Api.Services.interfaces;
 
 public interface ICurrentUser
 {
-    long Id { get; }
+    int Id { get; }
     string Username { get; }
     string Email { get; }
 }
